+++ Example/SequentDemo/SequentDemo/Sources/Features/Notes/NoteTransformer.swift
@@ -1,0 +1,16 @@
+import Brick
+
+struct NoteTransformer {
+
+  let note: Note
+
+  var listItem: Item {
+    return Item(
+      identifier: note.id,
+      title: note.title,
+      subtitle: "Author: \(note.userId)",
+      text: note.body,
+      kind: "list",
+      action: "notes:\(note.id)")
+  }
+}
